+++ src/main/java/com/sda/eventapp/web/mvc/form/validation/constraint/UniqueUsername.java
@@ -1,5 +1,6 @@
-package com.sda.eventapp.web.mvc.form.validation.email;
+package com.sda.eventapp.web.mvc.form.validation.constraint;
 
+import com.sda.eventapp.web.mvc.form.validation.validator.UniqueUsernameValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -10,9 +11,9 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = {UniqueEmailValidator.class})
-public @interface UniqueEmail {
-    String message() default "{com.sda.eventapp.web.mvc.form.validation.email.UniqueEmail.message}";
+@Constraint(validatedBy = {UniqueUsernameValidator.class})
+public @interface UniqueUsername {
+    String message() default "{com.sda.eventapp.web.mvc.form.validation.constraints.UniqueUsername.message}";
 
     Class<?>[] groups() default {};
