+++ src/main/java/com/sda/eventapp/web/mvc/form/validation/validator/UniqueEmailValidator.java
@@ -1,7 +1,8 @@
-package com.sda.eventapp.web.mvc.form.validation.username;
+package com.sda.eventapp.web.mvc.form.validation.validator;
 
 import com.sda.eventapp.service.UserService;
 import com.sda.eventapp.web.mvc.form.CreateUserForm;
+import com.sda.eventapp.web.mvc.form.validation.constraint.UniqueEmail;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
@@ -9,15 +10,15 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UniqueUsernameValidator implements ConstraintValidator<UniqueUsername, CreateUserForm> {
+public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, CreateUserForm> {
     private final UserService userService;
 
     @Override
-    public void initialize(UniqueUsername constraint) {
+    public void initialize(UniqueEmail constraint) {
     }
 
     @Override
     public boolean isValid(CreateUserForm form, ConstraintValidatorContext context) {
-        return !userService.existsByUsername(form.getUsername());
+        return !userService.existsByEmail(form.getEmail());
     }
 }
