+++ src/main/java/com/sda/eventapp/web/mvc/form/validation/validator/UniqueUsernameValidator.java
@@ -1,7 +1,8 @@
-package com.sda.eventapp.web.mvc.form.validation.email;
+package com.sda.eventapp.web.mvc.form.validation.validator;
 
 import com.sda.eventapp.service.UserService;
 import com.sda.eventapp.web.mvc.form.CreateUserForm;
+import com.sda.eventapp.web.mvc.form.validation.constraint.UniqueUsername;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.RequiredArgsConstructor;
@@ -9,15 +10,15 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, CreateUserForm> {
+public class UniqueUsernameValidator implements ConstraintValidator<UniqueUsername, CreateUserForm> {
     private final UserService userService;
 
     @Override
-    public void initialize(UniqueEmail constraint) {
+    public void initialize(UniqueUsername constraint) {
     }
 
     @Override
     public boolean isValid(CreateUserForm form, ConstraintValidatorContext context) {
-        return !userService.existsByEmail(form.getEmail());
+        return !userService.existsByUsername(form.getUsername());
     }
 }
